wrote all of the tables into the database.
Next step is to remove the chunks that completed these steps and to write a new table that has all three of the base tables joined.

I am going to test out a brute force join on a handful of columns to test the join and then will write the table... maybe a materizlied view? I need to see what dbos ( I know that dbo stands for data base owner, but I like to use dbos for database objects)